+++ test/fixtures/wrapper_fixtures/net_server_variation_1.js
@@ -1,0 +1,17 @@
+
+
+require('../../..').init({
+  ports : true,
+  network : true
+});
+
+var net = require('net');
+
+var server = net.createServer(function(socket) {
+  socket.on('data', function(data) {
+    socket.write(new Buffer(9999).fill('b') + '\r\n');
+  });
+});
+
+server.listen({ port : '9876' }, function() {
+});