+++ test/wrappers/http_wrapper.mocha.js
@@ -3,7 +3,6 @@ var pmx = require('../..');
 var request = require('request');
 var should = require('should');
 var fork = require('child_process').fork;
-var Plan = require('../helpers/plan');
 
 describe('HTTP Wrapper', function() {
   var app;
@@ -13,7 +12,7 @@ describe('HTTP Wrapper', function() {
   });
 
   it('should receive configuration flag', function(done) {
-    app = fork(__dirname + '/../fixtures/transaction/app.mock.auto.js', []);;
+    app = fork(__dirname + '/../fixtures/wrapper_fixtures/app.mock.auto.js', []);;
 
     app.once('message', function(data) {
       data.type.should.eql('axm:option:configuration');
@@ -22,69 +21,26 @@ describe('HTTP Wrapper', function() {
 
   });
 
-  it('should not log fast http request', function(done) {
-    var rcpt = function(data) {
-      if (data.type == 'axm:option:configuration')
-        return false;
-      if (data.type == 'axm:monitor')
-        return false;
+  it('should receive automatic custom metrics for HTTP server', function(done) {
+    var interval_req;
 
-      return data.type.should.not.eql('http:transaction');
+    var rcpt = function(packet) {
+      if (packet.type == 'axm:option:configuration')
+        return false;
+      if (packet.type == 'axm:monitor') {
+        should.exists((packet.data['HTTP'].value));
+        should.exists((packet.data['pmx:http:latency'].value));
+        app.removeListener('message', rcpt);
+        clearInterval(interval_req);
+        return done();
+      }
     };
 
     app.on('message', rcpt);
 
-    setTimeout(function() {
-      app.removeListener('message', rcpt);
-      return done();
-    }, 500);
-
-    setTimeout(function() {
+    interval_req = setTimeout(function() {
       request('http://127.0.0.1:9007/', function(req, res) {});
-    }, 100);
-  });
-
-  it('should not log ignored http request', function(done) {
-    var timer = setTimeout(function() {
-      app.removeListener('message', rcpt);
-      return done();
-    }, 1000);
-
-    var rcpt = function(data) {
-      if (data.type == 'axm:option:configuration')
-        return false;
-      if (data.type == 'axm:monitor')
-        return false;
-
-      return data.type.should.not.eql('http:transaction');
-    };
-
-    app.on('message', rcpt);
-
-    setTimeout(function() {
-      request('http://127.0.0.1:9007/socket.io/slow', function(req, res) {});
-    }, 100);
-  });
-
-  it('should log slow http request', function(done) {
-    var plan = new Plan(3, done);
-
-    app.on('message', function(data) {
-      if (data.type == 'axm:monitor') {
-        plan.ok(true);
-        if (Object.keys(data.data) < 3)
-          plan.ok(false);
-      }
-
-      if (data.type == 'http:transaction') {
-        data.data.should.have.properties('ip', 'time', 'url', 'method');
-        plan.ok(true);
-      }
-    });
-
-    setTimeout(function() {
-      request('http://127.0.0.1:9007/slow', function(req, res) {});
-    }, 100);
+    }, 10);
   });
 
 });