+++ CHANGELOG.md
@@ -12,6 +12,10 @@
 - Rename Loop Delay by Event Loop Latency
 - New method on pmx.init({ new_port : 20000 }) // it replaces the base port with the new one and broadcast to pm2 the base port ('lb:listening:port') // lib/wrapper/net_wrapper.js
 
+# 1.2.0
+
+- allow to pmx.emit without data
+
 # 1.1.0
 
 - new tracing system